Remove vchan connection data from xenstore on vchan close
When client find xenstore entries it tries to connect according to them,
and if something fails, whole client fails (in most cases). But when no
xenstore entries are found, client waits for it.
So to easy service restart (especially: qubesdb in case of DispVM)
clean old entries from xenstore on vchan close.

This needs to know xenstore path in libvchan_close, but unfortunately it isn't stored
in struct libxenvchan (nor data data needed to rebuild it). Because of
this we need to wrap struct libxenvchan in additional struct, which can
hold such data. This needs change in public header and every inline
function in it, which means need to recompile every vchan application.
So to prevent such problems in the future do not use inlines, even for
one-line wrapper functions.