client: watch for remote domain shutdown during client initialization
In libvchan_client_init we use xs.read_watch directly (no select loop),
so add a '@releaseDomain' watch to be notified when any domain is
destroyed, so we can check if our one is still alive.

This solves a problem with a lot of qubesdb-daemon processes in dom0
from failed VMs startups.