Check if remote domain is still alive
Resolve 'TODO' comment in libvchan_is_open. Also use the same code in
libvchan_client_init to check if server domain didn't died in the
meantime.

To perform the check xc_evtchn_status is used, but instead of checking
real event channel (which can be not connected yet), ask for status of
some (dummy) event channel of other domain. Xen performs a check for
that domain ID before validating event channel number, so we get ESRCH
error when domain ID is invalid (domain is dead). Otherwise we get
EINVAL because of port number is invalid. Theoretically we could also
get EPERM, because domU cannot check status of other domains ports.