Interface for async client connection
Currently only server connection can be established asynchronously. This
commit introduces proposed API for async client connection too.

Implementation-wise, the libvchan_client_init() consists mostly of
waiting-for-server code, and then calling libxenvchan_client_init().
libxenvchan_client_init() normally exits with an error if server is not
running yet. The idea is to split libvchan_client_init() into separate
functions that do not wait on its own, but delegate it to the
application as waiting for an FD to become readable (which will in
practice be waiting for a xenstore watch).

Add async connect interface, and then implement libvchan_client_init()
using this new interface.

Related to QubesOS/qubes-issues#7964